+++ src/webview/client/garage/app.js
@@ -11,7 +11,9 @@ const app = new Vue({
                 LABEL_DESC: 'Spawn your vehicles through this terminal.',
                 LABEL_MODEL: 'Model',
                 LABEL_FUEL: 'Fuel',
-                LABEL_CONTROLS: 'Controls'
+                LABEL_CONTROLS: 'Controls',
+                LABEL_SPAWN: 'spawn',
+                LABEL_DESPAWN: 'despawn'
             },
             vehicles: [],
             url: 'http://localhost:9111'