+++ documentation/changelog.md
@@ -44,6 +44,8 @@ description: List of all the changes and current versions.
 + Require Port 9111 for WebServer
 + Added Missing Locales for Chat Responses
 + Added Initial Garage Implementation
++ Added Garage Interface
++ Added Garage Vehicle Type Filtering
 ```
 
 ## 1.8.2 - April 19 - May 14, 2021